win_exec/win_exec.c: replace hash and key type about FindAPI.

+++ example/win_exec/win_exec.c
@@ -7,11 +7,11 @@ typedef uint (*WinExec)(LPCSTR lpCmdLine, uint uCmdShow);
 #pragma comment(linker, "/ENTRY:EntryMain")
 uint EntryMain() {
 #ifdef _WIN64
-    uint64 hash = 0x4BBDC64FE359FC5A;
-    uint64 key  = 0xD7D68112BE34E158;
+    uint hash = 0x4BBDC64FE359FC5A;
+    uint key  = 0xD7D68112BE34E158;
 #elif _WIN32
-    uint32 hash = 0x0AE20914;
-    uint32 key  = 0x61DA2999;
+    uint hash = 0x0AE20914;
+    uint key  = 0x61DA2999;
 #endif
     WinExec winExec = (WinExec)FindAPI(hash, key);
     if (winExec == NULL)
